backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1
asyncpg>=0.27.0
PyJWT>=2.8.0
